Update boxes-config with both f90 comments and box
The previous suggestion I made called `fortran` was for the box. However, older Fortran (i.e., FORTRAN 77) uses `c` as the comment character, not `!`. So, in case someone wants an `f77-box`, move the name to `f90-box`.

Also, a la the `ada-box` and `ada-cmt`, make a `f90-cmt` which does `!` commenting.