Fix static build on Linux with ncurses
When compiling our static ncurses, we need to have a valid TERMINFO variable, so that the fallback terminfo database can be retrieved. This must be a terminfo database in the newer, hashed format (hex directories), as required by ncurses >= 6.1. Normally, that's what you get when installing terminfo on Linux. On MinGW, it's also available, but sometimes in a weird directory.